#294 - Fix:코드 정리
